requirements.txt
Flask
line-bot-sdk
mecab-python3
unidic-lite
jaconv